Add support for debconf prompting to update-passwd
If DEBIAN_HAS_FRONTEND is set in the environment, update-passwd
will prompt with debconf for each change and skip changes that the
user declines.  Questions are generated dynamically for each
semantic change, instantiated from shared templates.  The only
exception are GECOS changes, where one question is used for all
GECOS changes for the same user.

update-passwd uses libdebconfclient to do the debconf prompting,
so add a build dependency on that package.

Modify postinst to use debconf prompting instead of manual
prompting if the debconf confmodule is available.